+++ app/Http/Controllers/Admin/VehicleController.php
@@ -14,7 +14,7 @@ class VehicleController extends Controller
      */
     public function index()
     {
-        Gate::authorize('index_vehicles_Vehicles');
+        Gate::authorize('index_vehicles_vehicles');
         $vehicles = Vehicle::latest()->get();
         return view('layouts.admin.vehicles.index', compact('vehicles'));
     }
@@ -25,11 +25,13 @@ class VehicleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create_vehicles_vehicles');
         return view('layouts.admin.vehicles.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create_vehicles_vehicles');
         $request->validate([
             'vehicle_name' => 'required',
             'brand' => 'required',
@@ -94,20 +96,23 @@ class VehicleController extends Controller
             ->with('success', 'Vehicle Created Successfully');
     }
 
-    public function edit(Vehicle $vehicle)
-    {
-        return view('layouts.admin.vehicles.create', compact('vehicle'));
-    }
-
     public function show(Vehicle $vehicle)
     {
+        Gate::authorize('view_vehicles_vehicles');
         $vehicle->load(['permits', 'services', 'repairs', 'tyreChanges']);
 
         return view('layouts.admin.vehicles.show', compact('vehicle'));
     }
 
+    public function edit(Vehicle $vehicle)
+    {
+        Gate::authorize('update_vehicles_vehicles');
+        return view('layouts.admin.vehicles.create', compact('vehicle'));
+    }
+
     public function update(Request $request, Vehicle $vehicle)
     {
+        Gate::authorize('update_vehicles_vehicles');
         $request->validate([
             'vehicle_name' => 'required',
             'brand' => 'required',
@@ -181,6 +186,7 @@ class VehicleController extends Controller
      */
     public function destroy(Vehicle $vehicle)
     {
+        Gate::authorize('delete_vehicles_vehicles');
         $vehicle->delete();
 
         return redirect()->route('admin.vehicles.index')
